вывел const в модуль
